models package refactor
