+++ src/main/java/com/configcat/ConfigCatClient.java
@@ -8,6 +8,10 @@ import com.configcat.hooks.ConfigCatHooks;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
 import com.configcat.log.LogLevel;
+import com.configcat.models.PercentageRule;
+import com.configcat.models.RolloutRule;
+import com.configcat.models.Setting;
+import com.configcat.models.SettingType;
 import com.configcat.override.OverrideBehaviour;
 import com.configcat.override.OverrideDataSource;
 import com.configcat.override.OverrideDataSourceBuilder;
