+++ src/main/java/com/configcat/evaluation/RolloutEvaluator.java
@@ -1,9 +1,11 @@
 package com.configcat.evaluation;
 
-import com.configcat.Setting;
 import com.configcat.User;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
+import com.configcat.models.PercentageRule;
+import com.configcat.models.RolloutRule;
+import com.configcat.models.Setting;
 import com.google.gson.JsonElement;
 import de.skuzzle.semantic.Version;
 import org.apache.commons.codec.binary.Hex;
