+++ src/main/java/com/configcat/fetch/ConfigFetcher.java
@@ -3,6 +3,8 @@ package com.configcat.fetch;
 import com.configcat.*;
 import com.configcat.log.ConfigCatLogMessages;
 import com.configcat.log.ConfigCatLogger;
+import com.configcat.models.Config;
+import com.configcat.models.Entry;
 import okhttp3.*;
 import org.jetbrains.annotations.NotNull;
