+++ src/main/java/com/configcat/models/Setting.java
@@ -1,7 +1,5 @@
-package com.configcat;
+package com.configcat.models;
 
-import com.configcat.evaluation.PercentageRule;
-import com.configcat.evaluation.RolloutRule;
 import com.google.gson.JsonElement;
 import com.google.gson.annotations.SerializedName;
